general tidy up, less variables in Dockerfile, and [[]] in init

+++ init/20_set_the_initial_config.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 mkdir -p /config/configs
-if [ ! -f "/config/znc.pem" ]; then
-/usr/local/bin/znc -d /config -p
-fi
+
+[[ ! -f /config/znc.pem ]] && /usr/local/bin/znc -d /config -p
+
 while [ ! -f "/config/znc.pem" ]; do
 echo "waiting for pem file to be generated"
 sleep 1s
 done
-if [ ! -f "/config/configs/znc.conf" ]; then
-cp /defaults/znc.conf /config/configs/znc.conf
-fi
+
+[[ ! -f /config/configs/znc.conf ]] && cp /defaults/znc.conf /config/configs/znc.conf
+
 chown -R abc:abc /config
+
